feat: change game-home to be game-play and create start menu
